[#1110] Extract document template concern from attestation template class
